+++ R/shinyRatings.R
@@ -2,7 +2,8 @@
 #'
 #' @param inputId The input slot that will be used to access the value of number of stars.
 #' @param no_of_stars Number of stars that you want to display on the UI. 
-#'
+#' @param default Number of stars selected by default.
+#' 
 #' @return Ratings to be added to UI definition
 #' @examples 
 #' if(interactive()){
@@ -23,9 +24,9 @@
 #' 
 #' @export
 #'
-shinyRatings <- function(inputId, no_of_stars = 5) {
-  stopifnot("no_of_stars can be a whole number or a number ending with .5" = no_of_stars %% 1 %in% c(0, .5))
-  
+shinyRatings <- function(inputId, no_of_stars = 5, default = no_of_stars) {
+  do_checks(no_of_stars, default)
+  calculate_def <- default * 2
   htmltools::tags$html(
     htmltools::tags$head(
       htmltools::tags$link(type = "text/css", href = "https://cdnjs.cloudflare.com/ajax/libs/font-awesome/4.7.0/css/font-awesome.min.css", rel = "stylesheet"),
@@ -34,7 +35,7 @@ shinyRatings <- function(inputId, no_of_stars = 5) {
     ),
     htmltools::tags$body(
       shiny::div(id = inputId, class = "shinyRatings", 
-                 htmltools::HTML(ratings_html(no_of_stars))
+                 htmltools::HTML(ratings_html(no_of_stars)), data = jsonlite::toJSON(list(n = calculate_def))
                 )
     )
   )
@@ -57,4 +58,12 @@ ratings_html <- function(n) {
     ), collapse = '')
   
   sprintf('<div class="rating-group">%s</div>', dynamic_html)
+}
+
+#' @noRd
+do_checks <- function(no_of_stars, default) {
+  # no_of_stars should be a whole number or should end with .5
+  stopifnot("no_of_stars can be a whole number or a number ending with .5" = no_of_stars %% 1 %in% c(0, .5))
+  stopifnot("default value can be a whole number or a number ending with .5" = default %% 1 %in% c(0, .5))
+  stopifnot("default value can be greater than 0 and less than `no_of_stars`" = default > 0 && default <= no_of_stars)
 }