+++ NEWS.md
@@ -1,3 +1,8 @@
+# shinyRatings 0.0.2
+
+* [Allow possibility of variable number of stars](https://github.com/shahronak47/shinyRatings/issues/9)
+* [Add default value of starts selected](https://github.com/shahronak47/shinyRatings/issues/9)
+
 # shinyRatings 0.0.1
 
 * [Fix center alignment of all the elements](https://github.com/shahronak47/shinyRatings/issues/11)